added support to do spbm_enum and fault. refoactor code to be more reusable
