+++ src/main/groovy/com/toastcoders/vmware/yavijava/parsers/DataObjectWSDLParserImpl.groovy
@@ -27,8 +27,13 @@ class DataObjectWSDLParserImpl implements WSDLParser {
 
     @Override
     void parse(String wsdl) {
+        parse(wsdl, [vim25: 'xmlns:vim25="urn:vim25"'])
+    }
+
+    @Override
+    void parse(String wsdl, Map nameSpace) {
         XmlSlurper slurper = new XmlSlurper()
-        def doc = slurper.parseText(wsdl).declareNamespace(['vim25': 'xmlns:vim25="urn:vim25"'])
+        def doc = slurper.parseText(wsdl).declareNamespace(nameSpace)
         String className = doc."@name"
         dataObject = new DataObject()
         dataObject.name = className
